Создал два модуля util.js, data.js. Перенес код в соответствующие модули.
